requirements.txt
attrs==21.2.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
PyYAML==6.0
toml==0.10.2
typing-extensions==4.0.1
